@@ -1,19 +1,22 @@
 import { Component, Input } from '@angular/core';
 import { Router } from '@angular/router';
+import { NoimagePipe } from '../../pipes/noimage.pipe';
 
 @Component({
   selector: 'app-card',
+  standalone: true,
+  imports: [NoimagePipe],
   templateUrl: './card.component.html',
-  styleUrls: ['./card.component.css']
+  styleUrls: ['./card.component.scss']
 })
 export class CardComponent {
 
   @Input() items: any[] = [];
   constructor(private router: Router) {}
 
-  verArtista(item: any) {
-    const artistaId = (item.type === 'artist') ? item.id : item.artists[0].id;
-    this.router.navigate(['/artist', artistaId]);
+  goToArtist(item: any) {
+    const artistId = (item.type === 'artist') ? item.id : item.artists[0].id;
+    this.router.navigate(['/artist', artistId]);
   }
 
 }
