@@ -1,9 +1,11 @@
 import { Component, OnInit } from '@angular/core';
+import { RouterModule } from '@angular/router';
 
 @Component({
   selector: 'app-navbar',
+  standalone: true,
+  imports: [RouterModule],
   templateUrl: './navbar.component.html',
-  styles: []
 })
 export class NavbarComponent implements OnInit {
 
