@@ -2,7 +2,8 @@ import { Pipe, PipeTransform } from '@angular/core';
 import { DomSanitizer } from '@angular/platform-browser';
 
 @Pipe({
-  name: 'domseguro'
+  name: 'domseguro',
+  standalone: true,
 })
 export class DomseguroPipe implements PipeTransform {
 
