@@ -1,7 +1,8 @@
 import { Pipe, PipeTransform } from '@angular/core';
 
 @Pipe({
-  name: 'noimage'
+  name: 'noimage',
+  standalone: true
 })
 export class NoimagePipe implements PipeTransform {
 
